backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
